# frontend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: MTN Bulk Data Wholesale Service API
  version: 0.1.0
  description: Backend contract required by the frontend-only MTN Bulk Data Wholesale Service console.
  x-api-requirements:
    - Every response must use the standard success/error envelope documented in components.schemas.ApiEnvelope and components.schemas.ErrorEnvelope.
    - Authenticated operations require the hardened HTTP-only JWT session cookie unless the operation explicitly sets security to an empty list.
    - Paginated table endpoints must return page, limit, total, totalPages, hasNextPage, and hasPreviousPage in meta.
    - Phone and MSISDN input must be normalized to E.164 Uganda numbers matching +256(77|78|79|76|39)XXXXXXX.
    - List endpoints with search parameters must support backend search because the frontend uses asynchronous API search when local table search has no matches.
    - Payment initiation must return a pending transaction and payment session that can be tracked globally until confirmation, expiry, or failure.
    - Customer account activation must follow the solution-document flow: activation link, email OTP verification, password creation, activation notification, then immediate portal access.
    - Bundle package validity is fixed at 30 days; configurable bundle fields are service code, name, volume up to 4 TB, price, status, and visibility.
...
servers:
  - url: https://backend.example.com/api
    description: Live backend base URL
security:
  - cookieAuth: []
...
paths:
...
  /purchases:
    post:
      tags: [Purchases]
      summary: Initiate bundle payment
      description: Creates a pending transaction and payment session. Mobile money payments use the configured MoMo provider; when SP transfer mode is enabled, the API posts FromFri, ToFri, Amount, Currency, ExternalTransactionId, and ReferenceId to ECW /sptransfer/ and waits for the asynchronous provider callback. Airtime payments use the update-balance-and-date integration at the configured 9040 base URL with originTimeStamp, originTransactionID, subscriberNumber, and a negative adjustmentAmountRelative, then wait for the asynchronous deduction callback. PRN payments call the configured Generate PRN provider endpoint with customerMsisdn, customerName, productCode, productName, amount, payingCurrency, transactionId, and additionalInfo, then return the provider paymentReference as the PRN while waiting for the paid callback; card payments initiate the configured card provider and return paymentUrl for frontend popup checkout. PCRF is called only after customer payment is confirmed, using the internal Provisioning subscribe-service adapter with msisdn, serviceCode, startDateTime, endDateTime, and transactionId. Subscribe timestamps use yyyy-MM-ddTHH:mm:ss. The frontend subscribes to the returned Socket.IO room/event for provider confirmation before provisioning.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseRequest"
            examples:
              prnPurchase:
                $ref: "#/components/examples/PurchaseRequestPrnExample"
              mobileMoneyPurchase:
                $ref: "#/components/examples/PurchaseRequestMobileMoneyExample"
              cardPurchase:
                $ref: "#/components/examples/PurchaseRequestCardExample"
      responses:
        "200":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseResponse"
              examples:
                prnPaymentInitiated:
                  $ref: "#/components/examples/PurchaseResponsePrnExample"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
  /purchases/{transactionId}/retry:
    post:
      tags: [Purchases]
      summary: Retry failed payment
      description: Creates a fresh payment session for a failed transaction and moves the transaction back to pending. Receipts must only be issued for provisioned transactions.
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseRetryRequest"
            examples:
              retryPrnPayment:
                $ref: "#/components/examples/PurchaseRetryRequestExample"
      responses:
        "200":
          description: Payment retry initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseResponse"
              examples:
                retryInitiated:
                  $ref: "#/components/examples/PurchaseRetryResponseExample"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
  /purchases/{transactionId}/confirmation:
    post:
      tags: [Purchases]
      summary: Confirm payment and provision bundle
      description: Confirms a provider payment event, provisions the bundle, and returns the Provisioning request flags. In production this is normally driven by the payment provider/webhook after a Socket.IO payment status event.
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseConfirmationRequest"
            examples:
              confirmPayment:
                $ref: "#/components/examples/PurchaseConfirmationRequestExample"
      responses:
        "200":
          description: Payment confirmed and bundle provisioned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseConfirmationResponse"
              examples:
                paymentConfirmed:
                  $ref: "#/components/examples/PurchaseConfirmationResponseExample"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
  /payments/callback:
    post:
      tags: [Payment Provider]
      summary: Receive payment provider callback
      description: Receives payment status callbacks from the external provider, including paid PRN, approved mobile money, and deducted airtime callbacks, provisions confirmed payments idempotently, and emits payment.status over Socket.IO.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentProviderCallbackRequest"
            examples:
              providerCallbackConfirmed:
                $ref: "#/components/examples/PaymentProviderCallbackExample"
      responses:
        "200":
          description: Callback processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseConfirmationResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
  /payments/provider-checkout/{sessionId}:
    get:
      tags: [Payment Provider]
      summary: Mock external provider checkout page
      description: Local development card checkout page. Real deployments should use the paymentUrl returned by the configured payment provider.
      security: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Mock checkout HTML
          content:
            text/html:
              schema:
                type: string
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /service-requests:
    get:
      tags: [Service Requests]
      summary: List service requests
      description: Admin report table for public onboarding requests with search, status, date range, pagination, and conversion workflow.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Search"
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [new, contacted, converted]
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
      responses:
        "200":
          description: Service requests fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequestListResponse"
              examples:
                serviceRequests:
                  $ref: "#/components/examples/ServiceRequestListResponseExample"
    post:
      tags: [Service Requests]
      summary: Submit public service request
      description: Captures a public lead from the MTN Bulk Data Wholesale public site for onboarding follow-up.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequestRequest"
            examples:
              publicServiceRequest:
                $ref: "#/components/examples/ServiceRequestRequestExample"
      responses:
        "200":
          description: Service request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequestResponse"
              examples:
                serviceRequestSubmitted:
                  $ref: "#/components/examples/ServiceRequestResponseExample"
        "422":
          $ref: "#/components/responses/ErrorResponse"
  /service-requests/{serviceRequestId}:
    parameters:
      - name: serviceRequestId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Service Requests]
      summary: Get service request
      responses:
        "200":
          description: Service request fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequestResponse"
              examples:
                serviceRequest:
                  $ref: "#/components/examples/ServiceRequestResponseExample"
        "404":
          $ref: "#/components/responses/ErrorResponse"
    patch:
      tags: [Service Requests]
      summary: Update service request status
      description: Lets admins mark public service requests as new or contacted before conversion.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequestUpdateRequest"
            examples:
              markContacted:
                $ref: "#/components/examples/ServiceRequestUpdateRequestExample"
      responses:
        "200":
          description: Service request updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequestResponse"
              examples:
                serviceRequestUpdated:
                  $ref: "#/components/examples/ServiceRequestUpdatedResponseExample"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
  /service-requests/{serviceRequestId}/convert:
    parameters:
      - name: serviceRequestId
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Service Requests]
      summary: Convert service request to customer
      description: Converts a qualified public service request into a customer profile using the supplied customer registration details.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRegistrationRequest"
            examples:
              convertServiceRequest:
                $ref: "#/components/examples/ServiceRequestConversionRequestExample"
      responses:
        "200":
          description: Service request converted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequestConversionResponse"
              examples:
                serviceRequestConverted:
                  $ref: "#/components/examples/ServiceRequestConversionResponseExample"
        "404":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          $ref: "#/components/responses/ErrorResponse"
        "422":
          $ref: "#/components/responses/ErrorResponse"
  /reports/admin:
    get:
      tags: [Reports]
      summary: Fetch admin reporting data
      description: Returns admin report datasets. Search, status, and date filters are used by frontend tables as an asynchronous fallback when local filtering finds no rows.
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: status
          in: query
          required: false
          description: Optional status filter applied to compatible report datasets.
          schema:
            type: string
            enum: [active, pending, deactivated, provisioned, failed]
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
      responses:
        "200":
          description: Admin report fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminReportResponse"
              examples:
                adminReport:
                  $ref: "#/components/examples/AdminReportResponseExample"
  /reports/transactions:
    get:
      tags: [Reports]
      summary: Fetch dedicated transaction report table
      description: Supports the frontend report table with search, customer, status, payment, date range, pagination, and infinite-scroll loading.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Search"
        - name: customerId
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [provisioned, pending, failed]
        - name: paymentMethod
          in: query
          required: false
          schema:
            type: string
            enum: [mobile_money, airtime, prn, card]
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
      responses:
        "200":
          description: Transaction report table fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportTransactionListResponse"
              examples:
                paginatedTransactions:
                  $ref: "#/components/examples/ReportTransactionListResponseExample"
  /reports/customer:
    get:
      tags: [Reports]
      summary: Fetch customer reporting data
      description: Returns customer report datasets. Search, status, and date filters are used by frontend tables as an asynchronous fallback when local filtering finds no rows.
      parameters:
        - name: customerId
          in: query
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/Search"
        - name: status
          in: query
          required: false
          description: Optional status filter applied to compatible bundle purchase and secondary number datasets.
          schema:
            type: string
            enum: [provisioned, pending, failed, active, removed]
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
      responses:
        "200":
          description: Customer report fetched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerReportResponse"
              examples:
                customerReport:
                  $ref: "#/components/examples/CustomerReportResponseExample"
        "404":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-mtn_bds_session
      description: HTTP-only, Secure, SameSite=Strict signed JWT session cookie. Local development may use mtn_bds_session over http.
  parameters:
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: One-based page number used by paginated tables.
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Number of rows per page. Infinite scrolling uses this as the fetch batch size.
    Search:
      name: search
      in: query
      required: false
      schema:
        type: string
      description: Case-insensitive search term applied to human-readable table fields.
...
    DateFrom:
      name: dateFrom
      in: query
      required: false
      schema:
        type: string
        format: date
      description: Inclusive start date for date-filtered table rows. For /overview, used when revenuePeriod is custom.
    DateTo:
      name: dateTo
      in: query
      required: false
      schema:
        type: string
        format: date
      description: Inclusive end date for date-filtered table rows. For /overview, used when revenuePeriod is custom.
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            validationError:
              $ref: "#/components/examples/ErrorResponseExample"
  examples:
...
    PurchaseRequestPrnExample:
      summary: Start PRN payment
      value:
        customerId: cus-wavenet
        primaryMsisdn: "+256772990001"
        bundleId: bundle-1tb
        provisioningCount: 2
        paymentMethod: prn
        prnProvider: bank
        additionalInfo: Customer requested weekend bundle
    PurchaseRequestMobileMoneyExample:
      summary: Start mobile money payment
      value:
        customerId: cus-wavenet
        primaryMsisdn: "+256772990001"
        bundleId: bundle-1tb
        provisioningCount: 1
        paymentMethod: mobile_money
        payingMsisdn: "+256772990001"
    PurchaseRequestCardExample:
      summary: Start card payment
      value:
        customerId: cus-wavenet
        primaryMsisdn: "+256772990001"
        bundleId: bundle-1tb
        provisioningCount: 1
        paymentMethod: card
        autoRenew: true
        redirectUrl: https://portal.example.com/payment-status
    PurchaseResponsePrnExample:
      summary: PRN payment initiated
      description: The PRN value is returned by the configured external PRN provider.
      value:
        success: true
        message: Payment initiated successfully
        data:
          transaction:
            id: txn-1009
            customerName: WaveNet
            primaryMsisdn: "+256772990001"
            bundleName: Wholesale 1 TB
            paymentMethod: prn
            amountUgx: 4600000
            status: pending
            createdAt: "2026-04-22T09:08:00+03:00"
          paymentSession:
            id: pay-20260422-001
            transactionId: txn-1009
            paymentMethod: prn
            status: awaiting_payment
            amountUgx: 4600000
            currency: UGX
            prn: A2086517260219
            provider: bank
            providerTransactionId: 6bc9766c-4f33-4493-a22c-1cc52c7ac365
            providerReference: A2086517260219
            providerGeneratedAt: "2026-02-19T11:30:00.000Z"
            socketEvent: payment.status
            socketRoom: payments:pay-20260422-001
            expiresAt: "2026-02-26T11:27:06.000Z"
            createdAt: "2026-04-22T09:08:00+03:00"
            customerId: cus-wavenet
            bundleId: bundle-1tb
            provisioningCount: 2
    PurchaseRetryRequestExample:
      summary: Retry failed transaction
      value:
        prnProvider: bank
    PurchaseRetryResponseExample:
      summary: Retry payment initiated
      value:
        success: true
        message: Payment retry initiated successfully
        data:
          transaction:
            id: txn-1012
            customerName: WaveNet
            primaryMsisdn: "+256772990001"
            bundleName: Wholesale 1 TB
            paymentMethod: prn
            amountUgx: 2300000
            status: pending
            createdAt: "2026-04-22T08:40:00+03:00"
          paymentSession:
            id: pay-20260422-002
            transactionId: txn-1012
            paymentMethod: prn
            status: awaiting_payment
            amountUgx: 2300000
            currency: UGX
            prn: PRN-20260422-0002
            provider: bank
            socketEvent: payment.status
            socketRoom: payments:pay-20260422-002
            expiresAt: "2026-04-22T10:05:00+03:00"
            createdAt: "2026-04-22T09:50:00+03:00"
            customerId: cus-wavenet
            bundleId: bundle-1tb
            provisioningCount: 1
    PurchaseConfirmationRequestExample:
      summary: Confirm payment session
      value:
        sessionId: pay-20260422-001
        status: confirmed
    PaymentProviderCallbackExample:
      summary: PRN provider confirms payment
      value:
        internalTransactionId: 6bc9766c-4f33-4493-a22c-1cc52c7ac365
        clientTransactionId: txn-1009
        paymentReference: A2086517260219
        status: paid
        receiptNumber: A2086517260219
    PurchaseConfirmationResponseExample:
      summary: Payment confirmed
      value:
        success: true
        message: Payment confirmed and bundle provisioned successfully
        data:
          transaction:
            id: txn-1009
            customerName: WaveNet
            primaryMsisdn: "+256772990001"
            bundleName: Wholesale 1 TB
            paymentMethod: prn
            amountUgx: 4600000
            status: provisioned
            createdAt: "2026-04-22T09:08:00+03:00"
          paymentSession:
            id: pay-20260422-001
            transactionId: txn-1009
            paymentMethod: prn
            status: confirmed
            amountUgx: 4600000
            currency: UGX
            prn: PRN-20260422-0001
            provider: bank
            socketEvent: payment.status
            socketRoom: payments:pay-20260422-001
            expiresAt: "2026-04-22T09:23:00+03:00"
            createdAt: "2026-04-22T09:08:00+03:00"
            customerId: cus-wavenet
            bundleId: bundle-1tb
            provisioningCount: 2
          provisioningRequest:
            subscribeService: true
            modifySubSubscription: true
            srvTopupCount: 1
    ServiceRequestRequestExample:
      summary: Public lead request
      value:
        businessName: WaveNet Uganda
        contactPerson: Sarah Namuli
        contactEmail: operations@wavenet.ug
        contactPhone: "+256772991001"
        preferredPackageId: bundle-1tb
        message: We need bulk data for managed branch devices.
...
  schemas:
...
    PurchaseRequest:
      type: object
      required:
        [customerId, primaryMsisdn, bundleId, provisioningCount, paymentMethod]
      properties:
        customerId:
          type: string
        primaryMsisdn:
          $ref: "#/components/schemas/UgandaPhoneNumber"
        bundleId:
          type: string
        provisioningCount:
          type: integer
          minimum: 1
          maximum: 12
        paymentMethod:
          type: string
          enum: [mobile_money, airtime, prn, card]
        payingMsisdn:
          $ref: "#/components/schemas/UgandaPhoneNumber"
        prnProvider:
          type: string
          enum: [bank, mobile_money]
          description: Required when paymentMethod is prn.
        autoRenew:
          type: boolean
        redirectUrl:
          type: string
          format: uri
          description: Card gateway return URL.
        additionalInfo:
          type: string
          maxLength: 300
          description: Optional PRN provider note.
    PurchaseRetryRequest:
      type: object
      properties:
        paymentMethod:
          type: string
          enum: [mobile_money, airtime, prn, card]
          description: Optional override. Defaults to the failed transaction payment method.
        payingMsisdn:
          $ref: "#/components/schemas/UgandaPhoneNumber"
        prnProvider:
          type: string
          enum: [bank, mobile_money]
          description: Required when retrying as PRN if the backend cannot infer a default provider.
        redirectUrl:
          type: string
          format: uri
          description: Card gateway return URL.
        additionalInfo:
          type: string
          maxLength: 300
          description: Optional PRN provider note.
...
    PurchaseConfirmationRequest:
      type: object
      required: [sessionId, status]
      properties:
        sessionId:
          type: string
        status:
          type: string
          enum: [confirmed, failed]
    PaymentProviderCallbackRequest:
      type: object
      description: >
        Provider callback. At least one identifier is required: sessionId,
        transactionId/clientTransactionId, internalTransactionId,
        providerReference, or paymentReference. PRN callbacks may omit status
        because a provider callback is treated as paid once the PRN session is
        resolved.
      properties:
        sessionId:
          type: string
        transactionId:
          type: string
        internalTransactionId:
          type: string
          example: 6bc9766c-4f33-4493-a22c-1cc52c7ac365
        clientTransactionId:
          type: string
          example: TXN-123456789
        paymentReference:
          type: string
          example: A2086517260219
        status:
          type: string
          enum:
            [
              awaiting_payment,
              processing,
              confirmed,
              paid,
              approved,
              deducted,
              failed,
              expired,
            ]
        providerReference:
          type: string
        receiptNumber:
          type: string
        failureReason:
          type: string
...
    ServiceRequestResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/ServiceRequest"
    ServiceRequestListResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [meta]
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/ServiceRequest"
            meta:
              $ref: "#/components/schemas/PaginationMeta"
...
    PurchaseResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/PurchaseResult"
    PurchaseConfirmationResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/PurchaseConfirmationResult"
...